mep_backend/requirements.txt
asgiref==3.5.2
dj-database-url==1.2.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.3.0
